Move plugins in Validation/RecoVertex into the plugins/ directory
