+++ Validation/RecoVertex/BuildFile.xml
@@ -5,31 +5,35 @@
 <use name="FWCore/ServiceRegistry"/>
 <use name="FWCore/Utilities"/>
 <use name="FWCore/Version"/>
+<use name="CommonTools/TriggerUtils"/>
+<use name="CommonTools/UtilAlgos"/>
 <use name="DataFormats/BeamSpot"/>
 <use name="DataFormats/Candidate"/>
+<use name="DataFormats/Common"/>
 <use name="DataFormats/Luminosity"/>
+<use name="DataFormats/Math"/>
 <use name="DataFormats/RecoCandidate"/>
 <use name="DataFormats/TrackReco"/>
 <use name="DataFormats/VertexReco"/>
+<use name="DPGAnalysis/SiStripTools"/>
 <use name="DQMServices/Core"/>
-<use name="SimDataFormats/Track"/>
-<use name="SimDataFormats/GeneratorProducts"/>
-<use name="SimDataFormats/PileupSummaryInfo"/>
 <use name="Geometry/TrackerGeometryBuilder"/>
 <use name="Geometry/Records"/>
 <use name="MagneticField/Records"/>
-<use name="SimDataFormats/Vertex"/>
+<use name="RecoVertex/PrimaryVertexProducer"/>
+<use name="RecoVertex/VertexTools"/>
+<use name="SimDataFormats/GeneratorProducts"/>
+<use name="SimDataFormats/PileupSummaryInfo"/>
+<use name="SimDataFormats/Track"/>
 <use name="SimDataFormats/TrackingAnalysis"/>
+<use name="SimDataFormats/Vertex"/>
 <use name="SimTracker/TrackerHitAssociation"/>
 <use name="SimTracker/TrackHistory"/>
 <use name="SimTracker/VertexAssociation"/>
-<use name="CommonTools/TriggerUtils"/>
-<use name="CommonTools/UtilAlgos"/>
 <use name="clhep"/>
-<use name="hepmc"/>
-<use name="RecoVertex/PrimaryVertexProducer"/>
-<use name="RecoVertex/VertexTools"/>
-<use name="root"/>
 <use name="gsl"/>
-<use name="DPGAnalysis/SiStripTools"/>
-<flags EDM_PLUGIN="1"/>
+<use name="hepmc"/>
+<use name="root"/>
+<export>
+  <lib name="ValidationRecoVertex"/>
+</export>
