+++ Validation/RecoVertex/plugins/BeamSpotAnalyzer.cc
