+++ Validation/RecoVertex/plugins/BuildFile.xml
@@ -1,0 +1,20 @@
+<use name="FWCore/Framework"/>
+<use name="FWCore/MessageLogger"/>
+<use name="FWCore/ParameterSet"/>
+<use name="FWCore/Utilities"/>
+<use name="CommonTools/UtilAlgos"/>
+<use name="DataFormats/Candidate"/>
+<use name="DataFormats/Common"/>
+<use name="DataFormats/VertexReco"/>
+<use name="DQMServices/Core"/>
+<use name="SimDataFormats/Associations"/>
+<use name="SimDataFormats/TrackingAnalysis"/>
+<use name="SimDataFormats/Vertex"/>
+<use name="SimTracker/VertexAssociation"/>
+<use name="Validation/RecoVertex"/>
+<lib name="MathMore"/>
+<use name="clhep"/>
+<use name="root"/>
+<library file="*.cc" name="ValidationRecoVertexPlugins">
+  <flags EDM_PLUGIN="1"/>
+</library>
